Base setup for views modified.
